Moved utility to correct folder
